+++ src/main/java/com/quasiris/qsf/query/parser/AbstractQueryParser.java
@@ -27,6 +27,9 @@ public abstract class AbstractQueryParser {
     }
 
     public void parse() {
+        if(tokens.isEmpty()) {
+            return;
+        }
 
         init();
         processNext();