+++ CHANGELOG.md
@@ -3,9 +3,23 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- implement a helper function in Document to get all values as a list
+
+### Removed
+
+### Changed
+
+### Deprecated
+
+### Fixed
+- AbstractQueryParser: don't parse the query if there are no tokens 
+
+## [2.10]
+### Added
 - implement a method to create a custom monitoring
 - implement addFilter method to simplify the creation of filter
 - implement a flag to deactive a monitoring
+- add the possibility to set a field and a direction in the sort of the search query
 
 ### Removed